swarm/storage: modularise storage package
