+++ swarm/storage/hasherstore/hasherstore_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package storage
+package hasherstore
 
 import (
 	"bytes"
