+++ swarm/storage/types.go
@@ -23,6 +23,7 @@ import (
 	"crypto/rand"
 	"encoding/binary"
 	"fmt"
+	"hash"
 	"io"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -324,4 +325,24 @@ func (f *FakeChunkStore) Get(_ context.Context, ref Address) (Chunk, error) {
 
 // Close doesn't store anything it is just here to implement ChunkStore
 func (f *FakeChunkStore) Close() {
+}
+
+const (
+	BMTHash     = "BMT"
+	SHA3Hash    = "SHA3" // http://golang.org/pkg/hash/#Hash
+	DefaultHash = BMTHash
+)
+
+type SwarmHash interface {
+	hash.Hash
+	ResetWithLength([]byte)
+}
+
+type HashWithLength struct {
+	hash.Hash
+}
+
+func (h *HashWithLength) ResetWithLength(length []byte) {
+	h.Reset()
+	h.Write(length)
 }
